requirements.txt
biopython
matplotlib
numpy
pandas
openmm
pdbfixer
mdtraj